beginning to work on the node layout

+++ buffer.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 
 /**
  * This represents a buffer with a size prefix.
@@ -8,7 +9,7 @@ struct buffer
     /**
      * This size of the buffer.
      */
-    int size;
+    uint16_t size;
 
     /**
      * The data payload with `size` number of bytes.
